Added popin action in accueil and started crawling tests
Added popin action for the elements with weather in the welcome page. Also
started to create the crawling tests. NB: one driver is created for each
tests. It takes about 8 seconds to launch. But when trying to use a class
variable instead, it throws an exception. Maybe I'll fix it later. Also,
keep the selenium-webdriver gem up to date or it won't work with FF (which
update continuously...)