Create the validate_joint_runner_function
Created the final runner validation function. Also, switching to a more
proactive commitng policy because I just lost 2 hours of work after glitch
while saving that caused the whole ts_TestCrawler.rb file to be erased.
Don't know if it's a Notepad++ once-in-a-decade bug or a hard drive
failure (more likely considering the HD I'm working on is not safely in
place) but I'll commit more to lose less if it happens again.