Polishes the tests after the refactoring
Changed the way the name is fetched to clear up the A that remained
sometimes. Also fixed the way the substitutes is interpreted, it broke when the way the name
was fetched changed. And added a filter on the runner lines: some (that
don't have the even or odd classes) are not runners but something else.