Adds a level of backup to tests.
After another Notepad++ crash that emptied ts_TestCrawler.rb (and deleted
a good hour and a half of progress along the way), I'm adding a backup
routine at the beginning of the test suite. The idea is simple: if hte
backup directory doesn't exist, it's created and everything (file and
folders) are copied within it. If it exists, only file that have been
modified since the last backup are saved. This comes in addition to the
backing up done through git and should reduce losses to five to ten (maybe
twenty in more complex cases) minutes of work lost.