Adds a comment to SimpleHtmlLogger
The comment is all that's left of an attempt to fail tests gracefully and
in a way that allowed tests to resume (after powering down the computer
for example). In the end, it's complex to do at the moment and I'm not
sure about the benefits (only ts_TestCrawler takes more than 5secs - it
takes a lot more (35 min atm) but the work is non-trivial.
In the end, I learned that you can't use mutexes in trap_context (the
context created on interruption).