Removed circular references
Removed circular references: Race in Runner and Meeting in Race. That way
no more infinite loops when loading a Runner or Race from the DB. Had to
redo almost everyting, which is why this is such a big commit.
Only Crawler and TestCrawler aren't compliant because there seems to be a
problem with the RefObjects. Which will be solved by more tests!