Finally corrected the random bug re: RefContainers
Tracked down after I remembered that the Tests are executed in random
order. Once that was figured out, it was simply an issue of tracking donw
how the fuck the RefContainer were getting duplicates. As it turned out,
when the RefContainers were re-created for each tests but the crawler was
not. And it was created with a reference to a ref_list_hash object (a
pointer to it). So the Crawler was adding Refs in one place and the tests
were adding them in another.
So the whole setup and teardown for the tests was regrouped in TestSuite
and the RefListHash was grouped in a static variable (rather than a
instance one, which was what got me one per test). So now, every one
shares the setup and teardown functions AND no more duplicates!