Finishes testFetchMeetings and fixes other tests.
Nothing big, testFetchMeetings seems to never have cbeen tested and needed
some kinks worked out. Other tests: in tsRef the special character test
lacked a differentiator so it failed if launched more that once. And there
was a weird bug with tsRef: loadAllRefs failed (with a ref missing when
checked against the database) but it cleared when I emptied the DB and
tried again. I hope it won't come back after another run of tsCrawler.