Added EndToEnd test, first integration test.
A simple integration test: fetch all meetings and save them. It's done so,
the first part of this app is FINALLY finished: a crawler that saves to a
RDB. Next up: setting it up on Heroku and attacking the real site to
retreive live data. (This will require adjustments is i'm sure the real
site changed since those tests pages were saved, 2 years ago.)